feature : 캘린더 설명 모달 구현 / 캘린더 권한 별 아이콘 재배치
