+++ component/friends/FriendsModal.tsx
@@ -1,33 +1,9 @@
 import { Modal } from "react-bootstrap";
-import styled from "styled-components";
+import { GreenCloseButton, CustomHeader, CustomBody, CustomFooter } from "../../styles/styledComponentModule";
 import FriendsList from "./FriendsList";
-
-const FriendsModalWrapper = styled.div``;
-
-const CustomHeader = styled(Modal.Header)`
-  border: none !important;
-`;
-const CustomBody = styled(Modal.Body)`
-  border: none;
-  overflow: auto;
-  max-height: 50rem;
-`;
-const CustomFooter = styled(Modal.Footer)`
-  border: none;
-`;
-
-
-export const GreenCloseButton = styled.div`
-  background-repeat: no-repeat;
-  background-position: center;
-  width: 2rem;
-  height: 2rem;
-  background-image: url("/assets/image/greenXCircle.svg"); ;
-`;
 
 const FriendsModal = (props) => {
   return (
-    <FriendsModalWrapper>
       <Modal
         {...props}
         size="lg"
@@ -45,7 +21,6 @@ const FriendsModal = (props) => {
         </CustomBody>
         <CustomFooter/>
       </Modal>
-    </FriendsModalWrapper>
   );
 };
