+++ component/share/TicketModal.tsx
@@ -2,19 +2,16 @@ import styled, { css } from "styled-components";
 import RedButton from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
 import ShareAPIButton from "./ShareAPIButton";
+import { CustomHeader } from "../../styles/styledComponentModule";
 
 const CenteredModalWrapper = styled.div`
   margin: auto;
 `;
 
-const CustomHeader = styled(Modal.Header)`
-  border: none !important;
-`;
-
 const CustomBody = styled(Modal.Body)`
   margin: 1rem 0rem 2rem 0rem;
   height: 20rem;
-  
+
   ${(props) =>
     props.background_img &&
     css`
@@ -23,7 +20,6 @@ const CustomBody = styled(Modal.Body)`
       background-position: center;
       background-size: contain;
     `}
-};
 `;
 
 const TabButton = styled(RedButton)`
@@ -32,19 +28,43 @@ const TabButton = styled(RedButton)`
   padding: 7px;
   color: white;
   background-size: 5px 5px;
-  background: rgba(255,255,255,.2);
+  background: rgba(255, 255, 255, 0.2);
   text-shadow: 1px 0 black;
   font-weight: bold;
   font-size: 24px;
   border: none !important;
-  background: linear-gradient(45deg, #AC473D 25%, #C0544A 0, #C0544A 50%, #AC473D 0, #AC473D 75%, #C0544A 0 );
-  &:hover{
-    background: linear-gradient(45deg, #3C6C54 25%, #78AB91 0, #78AB91 50%, #3C6C54 0, #3C6C54 75%, #78AB91 0 );
+  background: linear-gradient(
+    45deg,
+    #ac473d 25%,
+    #c0544a 0,
+    #c0544a 50%,
+    #ac473d 0,
+    #ac473d 75%,
+    #c0544a 0
+  );
+  &:hover {
+    background: linear-gradient(
+      45deg,
+      #3c6c54 25%,
+      #78ab91 0,
+      #78ab91 50%,
+      #3c6c54 0,
+      #3c6c54 75%,
+      #78ab91 0
+    );
   }
-  &:active{
-    background: linear-gradient(45deg, #3C6C54 25%, #78AB91 0, #78AB91 50%, #3C6C54 0, #3C6C54 75%, #78AB91 0 );
+  &:active {
+    background: linear-gradient(
+      45deg,
+      #3c6c54 25%,
+      #78ab91 0,
+      #78ab91 50%,
+      #3c6c54 0,
+      #3c6c54 75%,
+      #78ab91 0
+    );
   }
-`
+`;
 const TicketModal = (props) => {
   const onDownloadAs = (uri: string, filename: string) => {
     console.log("다운됨");
@@ -76,8 +96,12 @@ const TicketModal = (props) => {
           {props.text && <h2>{props.text}</h2>}
         </CustomBody>
         {/* ----------- footer ----------- */}
-        <ShareAPIButton/>
-        <TabButton onClick={()=>onDownloadAs(props.img, "pitapet_calendar.png")}>내 캘린더 저장하기</TabButton>
+        <ShareAPIButton />
+        <TabButton
+          onClick={() => onDownloadAs(props.img, "pitapet_calendar.png")}
+        >
+          내 캘린더 저장하기
+        </TabButton>
       </CenteredModalWrapper>
     </Modal>
   );
